prep test for game and separate functions from express route
